Ensure bindings gracefully shutdown dispatcher threads (#6)
* Do not initialize dispatching threads at construct time

Dispatching threads were stopping graceful shutdown. Instead of
initializing them when the object is constructed, they now
only listen when the object has connected. In order to ensure
they activate on the main thread, they are started and halted
in the Async worker complete callbacks.

The Activate and Deactivate methods can be called many times, as
they rely on checking the async member to test whether or not
they need to be activated or deactivated.

* Cleanup directory and fix deprecations