Refactor rebalance callback to more pure implementation (#31)
Rebalance callback can now be set like any other configuration value. If it is omitted, we rely on the internal librdkafka rebalance callback. If you set it to true, it will use a node implementation (and expose the event to you). If you provide a function, you will do the assignment yourself, and it will emit the event.