Removed useless intermediary class for Message
Intermediary class was doing little more than wrapping
the object so it could be checked for fatal errors.

Instead, consume can encapsulate message deleting
when it is an error and return a baton. Baton
has a valid message inside when it is not an error.

Then, we can leverage a converter which copies the data
over so the message can be safely deleted. This should
allow for graceful disconnects.