Closer 1:1 mapping of produce
1. Got rid of the original produce method and other
deprecated messages that no one was using.
2. RdKafka::err2str now has an exported library function
so the internal error class can use it to turn any old
integer into an error string. This is currently implemented
for the synchronous produce method, which now returns the same
int that the librdkafka return did.
3. Updated benchmark and it is showing around a 50% improvement
over the old async + object enumeration implementation