Change TopicWritable to a ProducerStream and TopicReadable to (#143)
KafkaConsumerStream

TopicWritable was limited in its use. It took producers and a topic name
and allowed for a buffer -> kafka interface for an individual topic.

I like the simplicity, but a lot of the logic ought to be reused for
a much more common scenario: producing to variable
topics/keys/partitions.

Enter ProducerStream. Run it in objectMode and you can produce objects.
If it isn't in object mode, provide it a topic and get the base
functionality for free.

It also is now decoupled from a "parent" producer. Rather, it is created
with a factory method so the producer inside is owned by the stream.
This means we can react to stream events like `end` appropriately,
Connecting and disconnecting from kafka is as easy as spinning up a
stream and getting rid of it when you're done.

Similarly, the Consumer stream is now decoupled from a "parent"
consumer. It is created similarly, using `createReadStream` as a factory
method on the Consumer class. Fewer changes were made to it, however
topics is now provided as an option on a stream object, and it now
supports non-objectMode, which will simply write the message buffer to
the stream.