Update dev dep: node-gyp to v5.x
Note that this is _not_ necessary for package _users_. Its a dev
dependency, but v5.x is the version that comes with current npm
releases, I suggest this package should track that. Note that 3.x is
quite old, and misses numerous bug fixes, so isn't going to work on lots
of systems, particularly Python 3 systems.