Reduce event listener type instantiation depth
The use of chained `extends` statements result in a very deep type instantiation for the listener function.
If the listener passed in also utilizes generics, Typescript will complain that `Type instantiation is excessively deep and possibly infinite.ts(2589)`
This can be reduced drastically by using a type map with the `keyof` keyword.