fix: release opaque data on error in produce
When there is an error trying to produce() messages,
no delivery report is ever generated for the message.
If the user specified `opaque` data for the message,
the call to produce() will has "pinned" it, but with
no delivery report, it is never released.  This change
checks for this condition and releases the opaque
data when this happens.

Fixes: #953