e2e/both: fix default consume timeout in reaching eof test
One round poll to broker is needed to read the message.
Second round poll to broker is needed to detect EOF.

The test is asserting that we get message and EOF within 2s.
But the consumer is configured with 'fetch.wait.max.ms': 1000.
We don't fit 2 timout out polls within 2s.

We need to increase the consume timeout to 3s.