e2e/both: fix default consume timeout in already at eof test
One round poll to broker is needed to read the message.
Second round poll to broker is needed to detect EOF.

The test is asserting that we get EOF, data and EOF
when data is produced at 2nd second.

But the consumer is configured with 'fetch.wait.max.ms': 1000.
We read the data in poll to broker slightly after 3rd second.
We detect the 2nd EOF slightly after 4th second.

We need to increase the consume timeout to 5s.