workers: add ability to set a total consume operation timeout to the consumenum worker
the consumenum worker consumes up to a certain # of messages, and only
applies the configured timeout to a per-message consumption, breaking
out of the consume loop when:

1. the end of the partition is reached (if partition eof messages are
enabled)
2. a consume times out
3. the batch size is filled

this leaves the worker susceptible to slow-loris type issues, where
messages come in at a cadence *just* under the timeout (e.g. every 900
milliseconds) -- with a batch size of 1000, that means that
`consumer.consume(1000, 1000, cb)` could take as many as 900 seconds
to complete, which is less than ideal.

introduce a configuration in the consumenum worker that allows callers
to specify a "total" timeout that gets applied to the entire operation
-- when exceeded, we break out of the loop and return the batch, no
matter how many messages were read.